partly done trying to fix inherit_doc for class hierarchies in python docs
